documentation and transformations updated
